fix taxt rate

+++ Delivery.Order.Domain/Factories/TaxFeeGenerator.cs
@@ -4,7 +4,7 @@ namespace Delivery.Order.Domain.Factories
 {
     public static class TaxFeeGenerator
     {
-        public static int GenerateTaxFees(int subTotalAmount, int taxRate)
+        public static int GenerateTaxFees(int subTotalAmount, decimal taxRate)
         {
             return (int)(Math.Round(((decimal)taxRate / 100) * subTotalAmount));
         }
